#1: Even bigger refactoring of how descriptors work. Split up into many classes with a common interface. Ready for testing!
